refactor : DTO 개념 적용 #9
