+++ YagomMarket/YagomMarket/Data/Network/DataMapping/ProductDeleteKeyRequestDTO.swift
@@ -1,0 +1,12 @@
+//
+//  ProductDeleteKeyRequestDTO.swift
+//  YagomMarket
+//
+//  Created by 이원빈 on 2023/01/20.
+//
+
+import Foundation
+
+struct ProductDeleteKeyRequestDTO: BodyType {
+    let secret: String
+}
