--- YagomMarket/YagomMarket/Domain/UseCases/Models/BodyType/DeleteKeyRequestModel.swift
@@ -1,12 +1,0 @@
-//
-//  DeleteKeyRequestModel.swift
-//  YagomMarket
-//
-//  Created by 이원빈 on 2022/11/16.
-//
-
-import Foundation
-
-struct DeleteKeyRequestModel: BodyType {
-    let secret: String
-}
